Revert "Use devm_device_add instead of sysfs_create"
This reverts commit 25d742bc3ef13d0e7042be7c74daf394da763999.

The proposed solution doesn't really work, there is no way to make sysfs
attributes static since (a) these attributes should be created if and
only if the proposed wmi guid is found so it's dynamic not static. (b)
adding groups to struct device_driver, in this case platform driver,
causes kernel oops! So making it static doesn't really make any sense
because not all devices have this guid that provides these
functionality.